refactor: rehome the Run aggregate — state machines to core, persistence to data
RunState/DimState machines now live in core/run/ (CONTEXT.md: Run is a
first-class domain concept); status.json/dimensions.json stores in data/fs.
run_lifecycle and provider prereqs move to analysis; shared keeps only the
node/npm checks. The project_resolver/repo_handler forwarders are deleted.
