+++ src/quodeq/core/run/__init__.py
@@ -1,0 +1,1 @@
+"""Run aggregate: lifecycle state machines for runs and dimensions."""
