--- src/quodeq/shared/repo_handler.py
@@ -1,4 +1,0 @@
-"""Re-export shim — canonical location is quodeq.data.fs.repo_handler."""
-from quodeq.data.fs.repo_handler import cleanup_cloned_repo, is_valid_repo_url, prepare_repository
-
-__all__ = ["cleanup_cloned_repo", "is_valid_repo_url", "prepare_repository"]
