+++ tests/analysis/test_cancellation.py
