+++ tests/api/test_evaluation_status_dim_states.py
@@ -10,7 +10,7 @@ import pytest
 from quodeq.api.app import create_app
 from quodeq.services.base import ActionProvider, EvaluationOptions
 from quodeq.services._job_model import JobSnapshot
-from quodeq.shared.dimensions_state import DimState, write_dim_state
+from quodeq.data.fs.dimensions_state_store import DimState, write_dim_state
 
 
 @pytest.fixture(autouse=True)
