+++ tests/ci/test_cli_lifecycle_integration.py
@@ -9,7 +9,7 @@ from unittest.mock import MagicMock, patch
 
 import pytest
 
-from quodeq.shared.run_status import read_status
+from quodeq.data.fs.run_status_store import read_status
 
 
 def test_pipeline_writes_running_then_done(tmp_path: Path) -> None:
@@ -91,7 +91,7 @@ def test_pipeline_writes_failed_on_domain_error(tmp_path: Path) -> None:
         result = cli._run_pipeline_with_cleanup(args, inputs, (tmp_path, evidence_dir, evaluation_dir))
         assert result == 1
 
-    from quodeq.shared.run_status import read_status
+    from quodeq.data.fs.run_status_store import read_status
     run_dir = evaluation_dir.parent
     status = read_status(run_dir)
     assert status["state"] == "failed"
@@ -106,7 +106,7 @@ def test_record_deadline_if_hit_tags_lifecycle_when_deadline_past(tmp_path: Path
     """_record_deadline_if_hit must call set_exit_reason('deadline') when
     config.options.deadline_at is in the past (i.e. loop broke on deadline)."""
     import quodeq._cli_evaluation as cli
-    from quodeq.shared.run_lifecycle import RunLifecycleContext
+    from quodeq.analysis.run_lifecycle import RunLifecycleContext
 
     run_dir = tmp_path / "run"
     run_dir.mkdir()
@@ -127,7 +127,7 @@ def test_record_deadline_if_hit_noop_when_no_deadline(tmp_path: Path) -> None:
     """If config.options.deadline_at is None, the helper must NOT touch
     exit_reason — a clean run still finalizes with exit_reason=null."""
     import quodeq._cli_evaluation as cli
-    from quodeq.shared.run_lifecycle import RunLifecycleContext
+    from quodeq.analysis.run_lifecycle import RunLifecycleContext
 
     run_dir = tmp_path / "run"
     run_dir.mkdir()
@@ -146,7 +146,7 @@ def test_record_deadline_if_hit_noop_when_deadline_not_yet_reached(tmp_path: Pat
     """If the deadline is still in the future when the loops returned (clean
     completion before the budget), the helper must NOT tag exit_reason."""
     import quodeq._cli_evaluation as cli
-    from quodeq.shared.run_lifecycle import RunLifecycleContext
+    from quodeq.analysis.run_lifecycle import RunLifecycleContext
 
     run_dir = tmp_path / "run"
     run_dir.mkdir()
@@ -239,7 +239,7 @@ def test_c88be50e_partial_state_invariants_agree(tmp_path: Path) -> None:
     import quodeq._cli_evaluation as cli
     from quodeq.analysis.cache.dimension_helpers import ClassifyResult
     from quodeq.analysis.cache.dimension_runner import _compute_files_read
-    from quodeq.shared.run_lifecycle import RunLifecycleContext
+    from quodeq.analysis.run_lifecycle import RunLifecycleContext
 
     # --- Half (a): lifecycle records exit_reason="deadline" -----------------
     run_dir = tmp_path / "run"
