+++ tests/services/test_evaluations_index_dedup.py
@@ -20,7 +20,7 @@ from pathlib import Path
 from quodeq.services._evaluations_index import EvaluationsIndex
 from quodeq.services._job_model import Job, InMemoryJobStore
 from quodeq.services.jobs import JobManager, STATUS_RUNNING
-from quodeq.shared.run_status import RunState, write_status
+from quodeq.data.fs.run_status_store import RunState, write_status
 
 
 def _seed_status(reports_root: Path, project: str, run_id: str) -> None:
