+++ tests/services/test_fs_projects.py
@@ -146,7 +146,7 @@ class TestUpdateProjectPath:
         assert info["path"] == str(new_dir.resolve())
         assert info["location"] == "local"
 
-    @patch("quodeq.shared.repo_handler.is_valid_repo_url", return_value=True)
+    @patch("quodeq.data.fs.repo_handler.is_valid_repo_url", return_value=True)
     def test_update_url_path(self, mock_valid, tmp_path: Path):
         reports_dir, project = self._setup_project(tmp_path)
         result = update_project_path(reports_dir, project, "https://github.com/org/repo.git")
@@ -154,7 +154,7 @@ class TestUpdateProjectPath:
         info = json.loads((Path(reports_dir) / project / "repository_info.json").read_text())
         assert info["location"] == "online"
 
-    @patch("quodeq.shared.repo_handler.is_valid_repo_url", return_value=False)
+    @patch("quodeq.data.fs.repo_handler.is_valid_repo_url", return_value=False)
     def test_rejects_invalid_url(self, mock_valid, tmp_path: Path):
         reports_dir, project = self._setup_project(tmp_path)
         assert update_project_path(reports_dir, project, "https://bad") is False
