+++ tools/import_baseline.txt
@@ -21,6 +21,3 @@ src/quodeq/services/evaluation_mixin.py:540:analysis
 src/quodeq/services/jobs.py:20:analysis
 src/quodeq/services/scan_progress.py:23:analysis
 src/quodeq/services/tooling_mixin.py:17:analysis
-src/quodeq/shared/prereqs.py:10:analysis
-src/quodeq/shared/project_resolver.py:2:data
-src/quodeq/shared/repo_handler.py:2:data
